move back generic and insiders dedicated sdk back to the sdk core to simplify development
