--- insiders/sdk/core/tests/__init__.py
@@ -1,1 +1,0 @@
-from tests.core import *
