--- insiders/sdk/core/tests/insiders/__init__.py
@@ -1,1 +1,0 @@
-from tests.core.universal_shipment import *
