--- insiders/sdk/generic/pyproject.toml
@@ -1,27 +1,0 @@
-[tool.poetry]
-name = "purplship.generic"
-version = "2021.12"
-homepage="https://next.purplship.com/docs/guides/sdk"
-repository="https://github.com/purplship/purplship"
-description = "Purplship - Generic Shipping Extension"
-authors = ["Purplship <hello@purplship.com>"]
-license = "Purplship Enterprise License"
-readme = "README.md"
-packages = [{ include = "purplship" }]
-classifiers=[
-    "Intended Audience :: Developers",
-    "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: Apache Software License",
-]
-
-[tool.poetry.dependencies]
-python = "^3.7"
-"purplship" = ""
-"purplship.insiders" = ""
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0", "setuptools!=50.0"]
-build-backend = "poetry.core.masonry.api"
