--- requirements.sdk.insiders.dev.txt
@@ -1,5 +1,0 @@
--r ./requirements.sdk.dev.txt
-
-# SDK extentions
--e ./insiders/sdk/core
--e ./insiders/sdk/generic
