+++ requirements.server.dev.txt
@@ -11,15 +11,12 @@ Django==3.2.9
 # -e ./sdk/extensions/dhl_express
 -e ./sdk/extensions/dhl_poland
 -e ./sdk/extensions/fedex
+-e ./sdk/extensions/generic
 # -e ./sdk/extensions/royalmail
 # -e ./sdk/extensions/sendle
 -e ./sdk/extensions/ups
 # -e ./sdk/extensions/usps
 # -e ./sdk/extensions/usps_international
-
-
--e ./insiders/sdk/core
--e ./insiders/sdk/generic
 
 
 purplship
