+++ scripts/setup-sdk-env.sh
@@ -5,10 +5,4 @@ source "scripts/create-new-env.sh"
 
 # Install requirements
 cd "${ROOT}" &&
-if [[ "$*" != *--insiders* ]];
-then
-    pip install -r "${ROOT:?}/requirements.sdk.dev.txt"
-else
-    pip install -r "${ROOT:?}/requirements.sdk.insiders.dev.txt"
-fi
-cd -
+pip install -r "${ROOT:?}/requirements.sdk.dev.txt"
