+++ sdk/core/tests/core/__init__.py
@@ -1,1 +1,2 @@
 from tests.core.universal_rate import *
+from tests.core.universal_shipment import *
