+++ sdk/extensions/generic/purplship/mappers/generic/settings.py
@@ -2,7 +2,7 @@
 
 from typing import List
 import attr
-from jstruct.types import JList, REQUIRED
+from jstruct.types import JList, REQUIRED, JStruct
 from purplship.core.models import ServiceLevel, LabelTemplate
 from purplship.universal.mappers.rating_proxy import RatingMixinSettings
 from purplship.universal.mappers.shipping_proxy import ShippingMixinSettings
@@ -14,12 +14,10 @@ from purplship.providers.generic.utils import Settings as BaseSettings
 class Settings(BaseSettings, RatingMixinSettings, ShippingMixinSettings):
     """Generic connection settings."""
 
-    account_number: str = None
-
     id: str = None
     test: bool = False
     carrier_id: str = "generic"
     account_country_code: str = None
 
+    label_template: LabelTemplate = JStruct[LabelTemplate]
     services: List[ServiceLevel] = JList[ServiceLevel, not REQUIRED, dict(default=DEFAULT_SERVICES)]  # type: ignore
-    templates: List[LabelTemplate] = JList[LabelTemplate]
