+++ sdk/extensions/generic/pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
-name = "purplship.insiders"
+name = "purplship.generic"
 version = "2021.12"
 homepage="https://next.purplship.com/docs/guides/sdk"
 repository="https://github.com/purplship/purplship"
-description = "Multi-carrier shipping API integration with python"
+description = "Purplship - Generic Shipping Extension"
 authors = ["Purplship <hello@purplship.com>"]
-license = "Apache-2.0"
+license = "Purplship Enterprise License"
 readme = "README.md"
 packages = [{ include = "purplship" }]
 classifiers=[
@@ -17,7 +17,7 @@ classifiers=[
 
 [tool.poetry.dependencies]
 python = "^3.7"
-purplship = ""
+"purplship" = ""
 
 [tool.poetry.dev-dependencies]
