+++ sdk/extensions/generic/setup.py
@@ -1,11 +1,12 @@
 """Warning: This setup.py is only there for git install until poetry support git subdirectory"""
+
 from setuptools import setup, find_namespace_packages
 
 setup(
-    name="purplship.insiders",
+    name="purplship.generic",
     version="0.0.0-dev",
-    license="Purplship Enterprise License",
-    packages=find_namespace_packages(exclude=["tests"]),
+    license="Apache-2.0",
+    packages=find_namespace_packages(),
     install_requires=["purplship"],
     zip_safe=False,
 )
