+++ sdk/extensions/generic/tests/generic/rate.py
